refactored star rendering and rewriting tests
