+++ client/src/components/RatingsAndReviews/Review.jsx
@@ -5,7 +5,7 @@ import moment from 'moment';
 // pass in a number to ratings for it to render a star
 
 const Review = ({ review }) => {
-  const { reviewer_name, date, summary, body, photos, recommend, response, helpfulness } = review;
+  const { reviewer_name, rating, date, summary, body, photos, recommend, response, helpfulness } = review;
   let photoRender;
   if (photos.length > 0) {
     photoRender = <li>
@@ -17,7 +17,7 @@ const Review = ({ review }) => {
   return (
     <div>
       <div className='reviewMeta'>
-        <Ratings />
+        <Ratings rating={rating}/>
         <p className='user'>{reviewer_name}, {moment(date).format('MMM Do YYYY')}</p>
       </div>
       <h4 className='reviewHeader'>{summary}</h4>
